Closes a few issues:
    Regression test (#46)
    Updated and extended documentation (#16, #55, #66)
    HW details for metrics and logging (#64)
    Tensorflow session builder (#64)
    Preprocessors short logging (#67)
    Metrics short logging (#68)
    Remove Docker -c option (#69)
    Model flexibility (#73)
    Dataset flexibility
    Fix M+ API MTE (#74)
    Better handling Kaggle credentials (#75)
    Unify preprocessors and transformers (#81)
    Closes #82 (outdated)
    Updated TensorFlow version (#88)
    Add data usage model to regression test (#54)

+++ src/lib/services/file_downloader.py
@@ -22,7 +22,21 @@ class FileDownloader:
     def __init__(self):
         self.logger = get_logger(prefix=self)
 
-    def download(self,  urls, save_dir=None, unzip=True, verbose=True):
+    def download(self, urls, save_dir=None, unzip=True, verbose=True):
+        """
+        Download file from URLs.
+
+        Parameters
+        ----------
+        urls: list or str
+            Urls to download.
+        save_dir: str, optional
+            Directory to save file. If unspecified, current working directory
+        unzip: bool, optional
+            Whether or not unzip downloaded files.
+        verbose: bool, optional
+            Whether or not show progress bar
+        """
         if save_dir is None:
             self.logger.debug("No saving directory provided. Saving to current directory")
             save_dir = os.getcwd()
@@ -32,10 +46,15 @@ class FileDownloader:
         else:
             self.__download_file(urls, save_dir, unzip=unzip, verbose=verbose)
 
-    def __download_file(self, url, save_dir, unzip=True, verbose=False):
+    def __download_file(self, url, save_dir, unzip=True, verbose=False, force=False):
         self.logger.info("Downloading file from url: {}".format(url))
         filename = os.path.basename(url)
         file_path = os.path.join(save_dir, filename)
+
+        if os.path.isfile(file_path) and not force:
+            self.logger.info("File {} already exists. Skipping".format(filename))
+            return
+
         with requests.get(url, stream=True) as r, open(file_path, "wb") as f:
             block_size = 1024
             total_size = int(r.headers.get('content-length', 0))
@@ -51,7 +70,7 @@ class FileDownloader:
 
         if can_be_decompressed(filename) and unzip:
             self.logger.info("Decompressing {}".format(file_path))
-            decompress(file_path)
+            decompress(file_path, delete_compressed=False)
 
 
     def __download_files(self, urls, save_dir, unzip=True, verbose=False):
